fixed clean_log_dir, refactored Vanilla,Classical,Aux loading with MNIST&CIFAR

+++ utils/logging.py
@@ -27,7 +27,7 @@ def allow_logging(func):
 
 
 def clean_log_dir(config, log_dir):
-    if config["clean_log_dir"] and len(os.listdir(log_dir)) > 0:  # Clears dir of old event files if True
+    if config["clean_log_dir"] and os.path.isdir(log_dir) and len(os.listdir(log_dir)) > 0:  # Clears dir of old event files if True
         for f in os.listdir(log_dir):
             try:
                 os.remove(os.path.join(log_dir, f))
